refactor bench files
