--- bench/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod cli;
-pub mod helpers;
-pub mod metrics;
